+++ BigBaseV2/src/util/toxic.hpp
@@ -30,7 +30,7 @@ namespace big::toxic
 	inline void bounty_player(Player target, int amount)
 	{
 		const size_t arg_count = 22;
-		int args[arg_count] = {
+		int64_t args[arg_count] = {
 			(int)eRemoteEvent::Bounty,
 			0, // doesn't matter of we set this to something else, the TRIGGER_SCRIPT_EVENT routine will set it to our player id anyways
 			target,
